+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - A StarfieldCustom.ini will be created with minimal values if is not already present. 
 - Symlink support has been re-enabled as this appears to work (it didn't in Fallout 4/Skyrim).
 - Add a warning when installing Starfield Script Extender on Xbox Game Pass. 
+- Fixed installing EXEs to the game root incorrectly. 
 
 ## [0.3.0] - 2023-09-17